requirements.txt
torch
torchvision
hydra-core
scikit-learn
numpy
matplotlib
seaborn